Former-commit-id: 6eedc6372af853ff5b03ffafd8469374467f7233

+++ Sources/Internal/Render/Image.cpp
@@ -181,7 +181,7 @@ void Image::ResizeCanvas(uint32 newWidth, uint32 newHeight)
 
 void Image::ResizeToSquare()
 {
-    float32 newImageSize = Max(width, height);
+    uint32 newImageSize = Max(width, height);
     ResizeCanvas(newImageSize, newImageSize);
 }
