+++ Tools/ResourceEditor/Classes/Qt/TextureBrowser/TextureCache.cpp
@@ -47,7 +47,7 @@ void TextureCache::setOriginal(const DAVA::TextureDescriptor *descriptor, DAVA::
 	if(NULL != descriptor)
 	{
 		DAVA::Vector<QImage> tmpImages;
-		for(int i = 0; i < images.size(); ++i)
+		for(size_t i = 0; i < images.size(); ++i)
 		{
 			tmpImages.push_back(images[i]);
 		}
